requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
streamlit==1.49.1
requests==2.32.5
pandas==2.3.2
numpy==2.1.3
scikit-learn==1.7.1
sentence-transformers==5.1.0
faiss-cpu==1.12.0
spacy==3.8.7
networkx==3.5
tqdm==4.67.1
pydantic==2.11.7
